requirements.txt
dwave-ocean-sdk>=6.10.0
django
djangorestframework